Pull from Features/clientcreate

+++ GoodSamaritan/Controllers/ClientController.cs
@@ -8,6 +8,8 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using GoodSamaritan.Models;
+using System.Diagnostics;
+using GoodSamaritan.Models.SmartEntity;
 
 namespace GoodSamaritan.Controllers
 {
@@ -19,6 +21,11 @@ namespace GoodSamaritan.Controllers
         // GET: Client
         public async Task<ActionResult> Index()
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
+
             var clientModel = db.ClientModel.Include(c => c.AbuserRealtionship).Include(c => c.Age).Include(c => c.AssignedWorker).Include(c => c.Crisis).Include(c => c.DuplicateFile).Include(c => c.Ethnicity).Include(c => c.FamilyViolenceFile).Include(c => c.FiscalYear).Include(c => c.Incident).Include(c => c.Program).Include(c => c.ReferralSource).Include(c => c.RepeatClient).Include(c => c.RiskLevel).Include(c => c.RiskStatus).Include(c => c.Service).Include(c => c.StatusOfFile).Include(c => c.VictimOfIncident);
             return View(await clientModel.ToListAsync());
         }
@@ -26,6 +33,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Details/5
         public async Task<ActionResult> Details(int? id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -41,6 +52,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Create
         public ActionResult Create()
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             ViewBag.AbuserRelationshipId = new SelectList(db.AbuserRelationshipModel, "AbuserRelationshipId", "AbuserRelationship");
             ViewBag.AgeId = new SelectList(db.AgeModel, "AgeId", "Age");
             ViewBag.AssignedWorkerId = new SelectList(db.AssignedWorkerModel, "AssignedWorkerId", "AssignedWorker");
@@ -68,11 +83,30 @@ namespace GoodSamaritan.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Create([Bind(Include = "ClientReferenceNumber,FiscalYearId,Month,Day,Surname,FirstName,PoliceFileNumber,CourtFileNumber,SwcFileNumber,RiskLevelId,CrisisId,ServiceId,ProgramId,RiskAssessmentAssignedTo,RiskStatusId,AssignedWorkerId,ReferralSourceId,IncidentId,AbuserSurnameName,AbuserFirstName,AbuserRelationshipId,VictimOfIncidentId,FamilyViolenceFileId,Gender,EthnicityId,AgeId,RepeatClientId,DuplicateFileId,NumberOfChildren0To6,NumberOfChildren7To12,NumberOfChildren13To18,StatusOfFileId,DateLastTransferred,DateClosed,DateReOpened")] ClientModel clientModel)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             if (ModelState.IsValid)
             {
+                var IsSmartModel = false;
+                if (clientModel.ProgramId == 3)
+                {
+                    IsSmartModel = true;
+                }
                 db.ClientModel.Add(clientModel);
                 await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+
+                if (IsSmartModel)
+                {
+                    SmartModel smart = new SmartModel() {
+                    };
+                    return RedirectToAction("Edit/", "Smart");
+                }
+                else
+                {
+                    return RedirectToAction("Index");
+                }
             }
 
             ViewBag.AbuserRelationshipId = new SelectList(db.AbuserRelationshipModel, "AbuserRelationshipId", "AbuserRelationship", clientModel.AbuserRelationshipId);
@@ -98,6 +132,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -134,6 +172,10 @@ namespace GoodSamaritan.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "ClientReferenceNumber,FiscalYearId,Month,Day,Surname,FirstName,PoliceFileNumber,CourtFileNumber,SwcFileNumber,RiskLevelId,CrisisId,ServiceId,ProgramId,RiskAssessmentAssignedTo,RiskStatusId,AssignedWorkerId,ReferralSourceId,IncidentId,AbuserSurnameName,AbuserFirstName,AbuserRelationshipId,VictimOfIncidentId,FamilyViolenceFileId,Gender,EthnicityId,AgeId,RepeatClientId,DuplicateFileId,NumberOfChildren0To6,NumberOfChildren7To12,NumberOfChildren13To18,StatusOfFileId,DateLastTransferred,DateClosed,DateReOpened")] ClientModel clientModel)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(clientModel).State = EntityState.Modified;
@@ -163,6 +205,10 @@ namespace GoodSamaritan.Controllers
         // GET: Client/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -180,6 +226,10 @@ namespace GoodSamaritan.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
+            if (!User.IsInRole("Worker") && !User.IsInRole("Administrator"))
+            {
+                return RedirectToAction("Denied", "Home");
+            }
             ClientModel clientModel = await db.ClientModel.FindAsync(id);
             db.ClientModel.Remove(clientModel);
             await db.SaveChangesAsync();
